libeos-updater-util: Split utilities out into separate noinst helper
Move the utility noinst helper library from within the src/ directory to
a new top-level directory, so that it’s more separate from the rest of
the code. This means it can be tested separately, and it’s harder to
introduce non-modular code into the project. The main motivation is to
separate testing of it, and to reduce the number of files in src/.

The files have been renamed to drop their ‘eos-’ prefix, but are
otherwise unchanged apart from #includes. The prefix is unnecessary
because the directory gives that. Compare:
   #include <libeos-updater-util/eos-extensions.h>
   #include <libeos-updater-util/extensions.h>
The prefix was also complicating tab-completion of filenames.

The build rules are mostly unchanged, but have been moved to the
top-level Makefile.am as the start of a slow migration to non-recursive
automake, which should speed builds up. This requires the subdir-objects
option in configure.ac, which has been supported by automake releases
for a long time now.

Overall, this should introduce no functional changes.

https://phabricator.endlessm.com/T15545
