+++ libeos-updater-util/extensions.h
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "eos-refcounted.h"
+#include <libeos-updater-util/refcounted.h>
 
 #include <ostree.h>
