+++ libeos-updater-util/refcounted.h
