+++ src/eos-updater-avahi.c
@@ -25,7 +25,7 @@
 #include "eos-updater-avahi-emulator.h"
 #include "eos-updater-avahi.h"
 
-#include "eos-util.h"
+#include <libeos-updater-util/util.h>
 
 #include "eos-updater-types.h"
