+++ src/eos-updater-fetch.c
@@ -24,7 +24,7 @@
 #include "eos-updater-fetch.h"
 #include "eos-updater-object.h"
 
-#include "eos-util.h"
+#include <libeos-updater-util/util.h>
 
 static void
 content_fetch_finished (GObject *object,
