+++ src/eos-updater-poll-common.c
@@ -25,7 +25,7 @@
 #include "eos-updater-object.h"
 #include "eos-updater-poll-common.h"
 
-#include "eos-util.h"
+#include <libeos-updater-util/util.h>
 
 #ifdef HAS_EOSMETRICS_0
