+++ src/eos-updater-poll-volume.c
@@ -25,7 +25,7 @@
 #include "eos-updater-poll-main.h"
 #include "eos-updater-poll-volume.h"
 
-#include "eos-util.h"
+#include <libeos-updater-util/util.h>
 
 #include <libsoup/soup.h>
